% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\title{Forecast Error of the Direct Forecast}
\usage{
error(df, lags, .H, .K, ic = BIC, .var = "all", .diff = FALSE,
  dev = "RMSE", wav = FALSE, xreg = NULL, ...)
}
\arguments{
\item{df}{A data frame. The first column of the data frame
must be the base variable for constructing the
dependent variable, the multiperiod ahead
value to be forecasted, \eqn{x_{t + h}}. If \code{wav = TRUE},
\code{ncol} of \code{df} can be equal to one (just the dependent variable),
otherwise, \code{ncol(df)} must be greater than one, even
if \code{.var = "ar"}.}

\item{lags}{An integer vector defining the lags of the
regressors. If \code{wav = FALSE}, the length of the vector \code{lags} have to
be the same as the number of columns in \code{df}. However,
if \code{wav = TRUE}, an additional element in \code{lags}
must be add such that the last element in the vector
\code{lags} is the lag associatd with the wavelet regressor.}

\item{.H}{An integer representing the maximum horizon step.}

\item{.K}{An integer, the number of pseudo-out-of-sample forecasts.}

\item{ic}{Information criterion, \code{BIC} or \code{AIC}. When searching
for the best model the dataset is adjusted so that every model
have the same data length
for appropriate comparasion.}

\item{.var}{A string to determine how the model will be
specificated: \code{"all"} (default), \code{"ar"} or \code{"ar_out"}.}

\item{.diff}{Logical \code{FALSE} or \code{TRUE}. If \code{TRUE}, the dependent
variable is differentiated. See \code{\link{wavdrcast-package}} for the
implication on the model specification and wavelet variable.}

\item{dev}{A string, \code{"RMSE"} or \code{"MAE"}.}

\item{wav}{Logical. If \code{TRUE}, a wavelet-based signal
is add to \code{df}, where the signal is related
to the explained variable, the first column of \code{df}.}

\item{xreg}{Data frame. Exogeunous variable not subjected
to be lagged. The number of rows must be the same as in \code{df}.}

\item{...}{Further arguments passed to \code{wavsigmap::\link[wavsigmap]{signal}}.}
}
\value{
A double vector with forecasts error (RMSE or MAE) from
1 to \code{.H} step-ahead.
}
\description{
Compute the root-mean-squared forecast error (RMSE) or
mean absolute forecast error (MAE) of direct forecast
OLS method for forecasting.
}
\details{
See \code{\link{wavdrcast-package}} and
\code{\link{model}}.
}
\examples{

cpi <- inf[1]
error(df = cpi,
lags = c(1, 1), .H = 2, .K = 8,
wav = TRUE, wt = "modwt", a = 1)

lgdp <- gdp[1]
error(df = lgdp,
lags = c(1, 1), .H = 2, .K = 8,
wav = TRUE, .diff = TRUE, wt = "modwt", a = 1)

}
\seealso{
\code{\link{model}, \link{fcast}}
}
