% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error_vin}
\alias{error_vin}
\title{Forecast Error of the Direct Forecast From Vintage}
\usage{
error_vin(.vin, lags, .H, .K, ic = BIC, .var = "all", .diff = FALSE,
  dev = "RMSE", wav = FALSE, xreg = NULL, ...)
}
\arguments{
\item{.vin}{List of data vintages (data frame) for each variable.
The first vintage is related to the dependent variable and
the number of columns of each data frame must be the
same as the number of the out-of-sample forecasts (\code{.K}).
The latest observation must be the last column of the data frame.
If \code{wav = TRUE}, the length of the list can be one,
otherwise, must be greater than one.}

\item{lags}{A integer vector defining the lags of the
regressors. If \code{wav = FALSE}, the length of the vector \code{lags} have to
be the same as the length of \code{.vin}. However,
if \code{wav = TRUE}, an additional element in \code{lags}
must be add such that the last element in the vector
\code{lags} is the lag of the wavelet regressor.}

\item{.H}{An integer representing the maximum horizon step.}

\item{.K}{An integer, the number of pseudo-out-of-sample forecasts.}

\item{ic}{Information criterion, \code{BIC} or \code{AIC}. When searching
for the best model the dataset is adjusted so that every model
have the same data length
for appropriate comparasion.}

\item{.var}{A string to determine how the model will be
specificated: \code{"all"} (default), \code{"ar"} or \code{"ar_out"}.}

\item{.diff}{Logical \code{FALSE} or \code{TRUE}. If \code{TRUE}, the dependent
variable is differentiated. See \code{\link{wavdrcast-package}} for the
implication on the model specification and wavelet variable.}

\item{dev}{A string, \code{"RMSE"} or \code{"MAE"}.}

\item{wav}{Logical. If \code{TRUE}, a wavelet-based signal
is add to \code{df}, where the signal is related
to the explained variable, the first column of \code{df}.}

\item{xreg}{Data frame. Exogeunous variable not subjected
to be lagged. The number of rows must be the same as in \code{df}.}

\item{...}{Further arguments passed to \code{wavsigmap::\link[wavsigmap]{signal}}.}
}
\value{
A double vector with forecasts error
(RMSE or MAE) from 1 to .H step-ahead.
}
\description{
Compute the root-mean-squared forecast error (RMSE) or
mean absolute forecast error (MAE) of the
direct forecast from vintage. This dataset is defined
as the sequence of values that represent the latest
estimate of the data at a particular moment in time.
}
\details{
See \code{\link{wavdrcast-package}} and
\code{\link{model}}.
}
\examples{
set.seed(1)

v1 <- tibble::tibble(v1_t1 = c(sample(100, 98, replace = TRUE), NA, NA),
                     v1_t2 = c(v1_t1[1:98], 500, NA),
                     v1_t3 = c(v1_t2[1:99], 280))

 set.seed(1)
v2 <- tibble::tibble(v2_t1 = 0.4 * v1[["v1_t1"]] + rnorm(100),
                     v2_t2 = c(v2_t1[1:98], 500 + rnorm(1), NA),
                     v2_t3 = c(v2_t2[1:99], 600 + rnorm(1)))

error_vin(list(v1, v2), lags = c(2, 2), .H = 3, .K = 3)
error_vin(list(v1), lags = c(2, 2), .H = 3, .K = 3, wav = TRUE)
}
\seealso{
\code{\link{model}, \link{fcast}, \link{error}}
}
