% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_best.R
\name{error_vin_wavmap}
\alias{error_vin_wavmap}
\title{Forecast Error Mapping Wavelet Models for Vintages}
\usage{
error_vin_wavmap(x, .vin, lags, .H, .K, ic = BIC, .var = "all",
  .diff = FALSE, dev = "RMSE", xreg = NULL, ...)
}
\arguments{
\item{x}{A tibble returned by the function
\code{wavsigmap::\link[wavsigmap]{map_wav_args}}.}

\item{.vin}{List of data vintages (data frame) for each variable.
The first vintage is related to the dependent variable and
the number of columns of each data frame must be the
same as the number of the out-of-sample forecasts (\code{.K}).
The latest observation must be the last column of the data frame.
If \code{wav = TRUE}, the length of the list can be one,
otherwise, must be greater than one.}

\item{lags}{A integer vector defining the lags of the
regressors. If \code{wav = FALSE}, the length of the vector \code{lags} have to
be the same as the length of \code{.vin}. However,
if \code{wav = TRUE}, an additional element in \code{lags}
must be add such that the last element in the vector
\code{lags} is the lag of the wavelet regressor.}

\item{.H}{An integer representing the maximum horizon step.}

\item{.K}{An integer, the number of pseudo-out-of-sample forecasts.}

\item{ic}{Information criterion, \code{BIC} or \code{AIC}. When searching
for the best model the dataset is adjusted so that every model
have the same data length
for appropriate comparasion.}

\item{.var}{A string to determine how the model will be
specificated: \code{"all"} (default), \code{"ar"} or \code{"ar_out"}.}

\item{.diff}{Logical \code{FALSE} or \code{TRUE}. If \code{TRUE}, the dependent
variable is differentiated. See \code{\link{wavdrcast-package}} for the
implication on the model specification and wavelet variable.}

\item{dev}{A string, \code{"RMSE"} or \code{"MAE"}.}

\item{xreg}{Data frame. Exogeunous variable not subjected
to be lagged. The number of rows must be the same as in \code{df}.}

\item{...}{Further arguments passed to
\code{wavsigmap::\link[wavsigmap]{signal}}.}
}
\value{
A tibble containing wavelet options and mean of
the forecast error.
}
\description{
\code{error_vin_wavmap} is a functional that computes the
root-mean-squared forecast error (RMSE) or mean absolute
forecast error (MAE) of the direct forecast for
 several wavelet models using data vintages.
}
\examples{
 set.seed(1)
v1 <- tibble::tibble(v1_t1 = c(sample(100, 98, replace = TRUE), NA, NA),
                     v1_t2 = c(v1_t1[1:98], 500, NA),
                     v1_t3 = c(v1_t2[1:99], 280))

wavmap <- wavsigmap::map_wav_args(list(
wavelet = c("haar", "s8"),
thresh.fun = c("universal",  "adaptive")
))

error_vin_wavmap(wavmap, .vin = list(v1),
lags = c(1, 1),
.H = 2, .K = 3)
}
\seealso{
\code{wavsigmap::\link[wavsigmap]{map_wav_args}},
\code{wavsigmap::\link[wavsigmap]{signal}}, \code{\link{error_vin}}
}
