% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Real Gross Domestic Product}
\format{
An object of class \code{tibble}. From 1949Qtr1 to 2019Qtr2.

\describe{
\item{lgdp}{\emph{Log of real gross domestic product}. \code{lgdp}
represents the natural logarithm of the real gross
domestic product, that is the inflation adjusted value of
the goods and services produced by labor and property
located in the United States.}

\item{gap}{\emph{Output gap}.
Defined as log(real gross domestic product) - log(potential gross domestic product). Real potential
GDP is the CBO’s estimate of the output the
economy would produce with a high rate of use of its
capital and labor resources. The data is adjusted to
remove the effects of inflation.}
}
}
\source{
Series can be obtained  in: \cr
\href{https://fred.stlouisfed.org/graph/?id=GDPC1,GDPPOT}{Federal Reserve Economic Data | FRED | St. Louis Fed}
}
\usage{
gdp
}
\description{
Real gross domestic product is the inflation adjusted value of
the goods and services produced by labor and property located in the United States.
}
\references{
U.S. Bureau of Economic Analysis,
Real Gross Domestic Product [GDPC1], retrieved from FRED,
\href{https://fred.stlouisfed.org/series/GDPC1}{Federal Reserve Bank of St. Louis}, September 11, 2019.

U.S. Congressional Budget Office, Real Potential
Gross Domestic Product [GDPPOT], retrieved from FRED, Federal
\href{https://fred.stlouisfed.org/series/GDPPOT}{Reserve Bank of St. Louis}, September 11, 2019.
}
\keyword{datasets}
