% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inf}
\alias{inf}
\title{US Inflation and Core Inflation Data}
\format{An object of class (inherits from tbl, data.frame) with 598 rows and 2 columns.
Monthly \% variation from 1970-01-01 to 2019-10-01.

\describe{
\item{cpi}{\emph{Consumer Price Index for All Urban Consumers: All Items in U.S. City Average  (CPIAUCNS)}. \code{cpi}
represents the variation in consumer price index. Based on Index 1982-1984=100,
not seasonally adjusted}

\item{core}{\emph{Consumer Price Index for All Urban Consumers: All Items Less Food and Energy in U.S. City Average  (CPILFENS)}.
Excludes food and energy. Based on Index 1982-1984=100, not seasonally adjusted}
}}
\source{
Series can be obtained  in: \cr
\href{https://fred.stlouisfed.org/graph/?id=CPIAUCNS,CPILFENS}{Federal Reserve Economic Data | FRED | St. Louis Fed}
}
\usage{
inf
}
\description{
A dataset containing the variation of the US headline Consumer Price Index
along with a core inflation measure that exclude food and energy.
}
\references{
U.S. Bureau of Labor Statistics,
Consumer Price Index for All Urban Consumers:
All Items Less Food and Energy in U.S.
City Average [CPILFENS], retrieved from FRED,
\href{https://fred.stlouisfed.org/series/CPIAUCNS}{Federal Reserve Bank of St. Louis}, November 26, 2019.

U.S. Bureau of Labor Statistics,
Consumer Price Index for All Urban Consumers:
All Items Less Food and Energy in U.S.
City Average [CPILFENS], retrieved from FRED,
\href{https://fred.stlouisfed.org/series/CPILFENS}{Federal Reserve Bank of St. Louis}, November 26, 2019.
}
\keyword{datasets}
