% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Direct Model}
\usage{
model(
  df,
  lags,
  .h,
  ic = BIC,
  .var = "all",
  .diff = FALSE,
  wav = FALSE,
  xreg = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame. The first column of the data frame
must be the base variable for constructing the
dependent variable, the multiperiod ahead
value to be forecasted, \eqn{x_{t + h}}. If \code{wav = TRUE},
\code{ncol} of \code{df} can be equal to one (just the dependent variable),
otherwise, \code{ncol(df)} must be greater than one, even
if \code{.var = "ar"}.}

\item{lags}{An integer vector defining the lags of the
regressors. If \code{wav = FALSE}, the length of the vector \code{lags} have to
be the same as the number of columns in \code{df}. However,
if \code{wav = TRUE}, an additional element in \code{lags}
must be add such that the last element in the vector
\code{lags} is the lag associatd with the wavelet regressor.}

\item{.h}{\eqn{h} step-ahead.}

\item{ic}{Information criterion, \code{BIC} or \code{AIC}. When searching
for the best model the dataset is adjusted so that every model
have the same data length
for appropriate comparasion.}

\item{.var}{A string to determine how the model will be
specificated: \code{"all"} (default), \code{"ar"} or \code{"ar_out"}.}

\item{.diff}{Logical \code{FALSE} or \code{TRUE}. If \code{TRUE}, the dependent
variable is differentiated. See \code{\link{wavdrcast-package}} for the
implication on the model specification and wavelet variable.}

\item{wav}{Logical. If \code{TRUE}, a wavelet-based signal
is add to \code{df}, where the signal is related
to the explained variable, the first column of \code{df}.}

\item{xreg}{Data frame. Exogeunous variable not subjected
to be lagged. The number of rows must be the same as in \code{df}.}

\item{...}{Further arguments passed to \code{wavsigmap::\link[wavsigmap]{signal}}.}
}
\value{
A list with the following elements:
\describe{
\item{xreg}{A tibble with the data regressors used to run the
model choosed by the information criterion. This can be
a different data set from that one returned by the \code{lm} object.}
\item{model}{An object of class \code{lm}.}
}
}
\description{
This function estimates a direct least square
horizon-specific model.
}
\details{
General OLS regression model is writting as:

    \eqn{x_{t + h} = \alpha + \mathbf{{\beta}}'\mathbf{x}_t
    + \mathbf{{\gamma}}'\mathbf{z}_t +
    \mathbf{{\theta}}'\mathbf{d}_t + \epsilon_{t + h}},

where variables are explained in \code{\link{wavdrcast-package}} details
section. Arguments \code{.var, .diff} and \code{wav} allow one to
estimate subsets of this general model,
as discussed in \code{\link{wavdrcast-package}}. If
\code{.var = "all"}, the general model is estimated (unless
\code{xreg = NULL}, so that \eqn{\mathbf{d}} is out). If
\code{.var = "ar"}, \eqn{\mathbf{z}} terms are excluded
(not \eqn{\mathbf{d}}, if included). Finally, if
\code{.var = "ar_out"}, the model is estimated without
\eqn{\mathbf{x}} regressors.
}
\examples{
model(df = inf, lags = c(2, 0), .h = 2, .var = "ar")
model(df = inf, lags = c(2, 2), .h = 2)
model(df = inf[1], lags = c(2, 2), .h = 2, wav = TRUE)
model(df = gdp, lags = c(2, 2), .h = 2, .diff = TRUE)
model(df = gdp[1], lags = c(2, 2), .h = 2, .diff = TRUE, wav = TRUE)
model(df = gdp[1], lags = c(0, 0), .h = 2, .diff = TRUE, wav = TRUE, .var = "ar_out")

}
\seealso{
\code{\link{fcast}}
}
