% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_a}
\alias{prior_a}
\title{Determine Scale Factor Paramenter "a" for Laplace Distribution}
\usage{
prior_a(..., wt = "modwt", interval = c(0.1, 5), tol = 0.01)
}
\arguments{
\item{...}{Parameters for function \code{\link{error}}.
Argument \code{wav = TRUE} is automatically adjusted.}

\item{wt}{A string "modwt" (maximal overlap
discrete wavelet transform) or
"dwt" (discrete wavelet transform).}

\item{interval}{A vector containing the end-points of the interval to be searched for the minimum. See: \code{stats::\link[stats]{optimize}}.}

\item{tol}{The desired accuracy. See: \code{stats::\link[stats]{optimize}}.}
}
\value{
A one-dimension double vector indicating the \code{a} prior.
}
\description{
The function prior_a determines the parameter a
of the laplace distribution that is used in
\href{https://www.rdocumentation.org/packages/EbayesThresh/versions/1.4-12/topics/ebayesthresh.wavelet}{empirical
Bayes thresholding approach}. Estimation is based on
the avarage of the forecast
error.
}
\examples{
prior_a(df = gdp[1],
        lags = c(1, 1), .H = 2, .K = 8,
        .diff = TRUE, .var = "ar_out", vscale = "level")
}
