% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_alpha_beta}
\alias{prior_alpha_beta}
\title{Determine Hyperparameters for Bayesian Wavelet Thresholding}
\usage{
prior_alpha_beta(
  ...,
  par = c(0.5, 1),
  lower = c(0, 0),
  upper = c(3, 3),
  control = list()
)
}
\arguments{
\item{...}{Parameters for function error. Argument wav = TRUE is automatically adjusted.}

\item{par}{Initial values for the parameters to be optimized over.}

\item{lower}{Bounds on the variables for the "L-BFGS-B" method. See \code{stats::\link[stats]{optim}}.}

\item{upper}{Bounds on the variables for the "L-BFGS-B" method. See  \code{stats::\link[stats]{optim}}.}

\item{control}{A list of control parameters. See \code{stats::\link[stats]{optim}}.}
}
\value{
A double vector of length one.
}
\description{
The function \code{prior_alpha_beta} determines both the parameters
\code{alpha} and \code{beta} that are used
in  \href{https://www.rdocumentation.org/packages/wavethresh/versions/4.6.1/topics/BAYES.THR}{bayesian wavelet thresholding
of noisy data}. Estimation is based on the avarage of the forecast error.
}
\examples{
prior_alpha_beta(df = inf[1],
lags = c(1, 1), .H = 2, .K = 2,
type = "hard", boundary = FALSE)
}
