% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vint.R
\name{wavint}
\alias{wavint}
\title{Wavelet Vintage}
\usage{
wavint(x, k, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{Number of vintages}

\item{...}{Addition parameters for wavsigmap::signal}
}
\value{
A tibble
}
\description{
Wavelet Vintage
}
\examples{
lgdp <- gdp[[1]]

vin_wavgap <- wavint(lgdp, 24,
                     wf = "haar",
                     vscale = "level", a = 1,
                     threshrule = "hard"
) \%>\%
  dplyr::mutate_all(~ `-`(lgdp, .))

}
